разделил модели по отдельным файлам
