+++ src/models/genre.py
@@ -1,0 +1,10 @@
+from models.film import BaseModelOrjson
+
+
+class Genres(BaseModelOrjson):
+    """
+    Схемы ответов для:
+    /api/v1/genres/
+    /api/v1/genres/<uuid:UUID>/
+    """
+    name: str
